fix(core-ui): keep the global search shortcut alive without the tabs row
The tabs row renders nothing while hidden, which unmounted the search
dialog along with its keydown listener, so the shortcut only worked with
the row on screen. The dialog and the shortcut now live in the layout and
the row keeps a trigger that flips the shared open state.

Also key the highlighted match on its offset instead of the array index,
and escape the searched term with String.raw.
