+++ frontend/core-ui/src/modules/navigation/components/VisitedPageTabs.tsx
@@ -1,4 +1,4 @@
-import { GlobalSearch } from '@/search/components/GlobalSearch';
+import { GlobalSearchTrigger } from '@/search/components/GlobalSearchTrigger';
 import { VisitedPageTabsShortcutGuide } from '@/navigation/components/VisitedPageTabsShortcutGuide';
 import { useNavigationActivities } from '@/navigation/hooks/useNavigationActivities';
 import { usePluginsModules } from '@/navigation/hooks/usePluginsModules';
@@ -418,7 +418,7 @@ export const VisitedPageTabs = () => {
           {tabItems}
         </VisitedPageTabsContent>
       </div>
-      <GlobalSearch />
+      <GlobalSearchTrigger />
       <Button
         aria-keyshortcuts={toggleTabsAriaShortcut}
         aria-label={t('navigation.hide-tabs-row')}
