+++ frontend/core-ui/src/modules/search/states/globalSearchState.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai';
+
+export const globalSearchOpenState = atom<boolean>(false);
