+++ frontend/core-ui/src/modules/search/utils/highlightMatch.tsx
@@ -1,5 +1,5 @@
 const escapeRegExp = (value: string) =>
-  value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+  value.replace(/[.*+?^${}()|[\]\\]/g, String.raw`\$&`);
 
 // Splits the text on the searched term and marks the matches. The API decides
 // what matched, so a hit on a field the row does not render simply comes back
@@ -16,17 +16,23 @@ export const highlightMatch = (
 
   const parts = text.split(new RegExp(`(${escapeRegExp(term)})`, 'ig'));
   const lowerTerm = term.toLowerCase();
+  let offset = 0;
 
-  return parts.map((part, index) =>
-    part.toLowerCase() === lowerTerm ? (
+  return parts.map((part) => {
+    const start = offset;
+    offset += part.length;
+
+    if (part.toLowerCase() !== lowerTerm) {
+      return part;
+    }
+
+    return (
       <mark
-        key={index}
+        key={`${start}:${part}`}
         className="rounded-[2px] bg-yellow-100 text-foreground dark:bg-yellow-500/25"
       >
         {part}
       </mark>
-    ) : (
-      part
-    ),
-  );
+    );
+  });
 };
